pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "duck-egg-fertility-detection"
version = "1.0.0"
description = "Duck Egg Fertility Detection using Deep Learning and Clustering"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
requires-python = ">=3.10"
dependencies = [
    # Core ML dependencies (add from requirements.txt)
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",
    "opencv-python>=4.8.0",
    " pillow>=10.0.0",
    "PyYAML>=6.0",
    
    # Deep Learning
    "tensorflow>=2.13.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    
    # Web Framework
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "streamlit>=1.28.0",
    
    # Database
    "sqlalchemy>=2.0.0",
    
    # Visualization
    "matplotlib>=3.8.0",
    "seaborn>=0.12.0",
    "plotly>=5.18.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.6.0",
    "ipython>=8.18.0",
]

docs = [
    "sphinx>=7.2.0",
    "sphinx-rtd-theme>=1.3.0",
]

[project.scripts]
duck-egg-train = "scripts.03_train_unet:main"
duck-egg-predict = "scripts.07_inference:main"
duck-egg-serve = "web.api.app:main"

[project.urls]
Homepage = "https://github.com/yumikopubangelo/duck_egg_fertility_detection"
Repository = "https://github.com/yumikopubangelo/duck_egg_fertility_detection"
Issues = "https://github.com/yumikopubangelo/duck_egg_fertility_detection/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "app*", "web*"]
exclude = ["tests*", "docs*", "deployment*", "notebooks*"]

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
    \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | venv
    | _build
    | buck-out
    | build
    | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long
    "B008", # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["B008"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src/ --cov-report=term-missing"

[tool.coverage.run]
source = ["src", "app", "web"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]
